python/pyproject.toml
[project]
name = "cpp-twocrypto-benchmark"
version = "0.1.0"
description = "Benchmark comparison between C++ and Vyper StableswapMath implementations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "titanoboa>=0.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "plotly>=5.0.0",
    "snekmate==0.1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ipython>=8.0.0",
]
